wip(justin-cli): stdin mocking for inquirer prompt testing
Kind of got the mocking in the right order I expect, with a _simple_ test case.
Have to cleanup, extract into a test utility, test with the full action.
Have to also check with the console output / stdout capturing util that the prompt works with that capturing output (should do).
And silence the output when we want.